fix: change DividendYear from int to decimal to match database
- change FundDiv.DividendYear from int? to decimal? to match DECIMAL(4,0) in database
- update EF configuration to specify decimal(4,0) type for DividendYear
- fixes InvalidCastException when querying FundDiv data

+++ DmsSystem.Domain/Entities/FundDiv.cs
@@ -13,7 +13,7 @@ public partial class FundDiv
     /// <summary>
     /// 配息年度
     /// </summary>
-    public int? DividendYear { get; set; }
+    public decimal? DividendYear { get; set; }
 
     /// <summary>
     /// 配息基準日
